// src/types/generated/productResponseDto.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * Inventory Management API
 * A comprehensive NestJS inventory and order management system with role-based access control, stock tracking, and audit logging.
 * OpenAPI spec version: 1.0
 */
import type { BrandResponseDto } from './brandResponseDto';
import type { CategoryResponseDto } from './categoryResponseDto';
import type { ManufacturerResponseDto } from './manufacturerResponseDto';
import type { ProductInventoryDto } from './productInventoryDto';
import type { ProductResponseDtoBarcode } from './productResponseDtoBarcode';
import type { ProductResponseDtoDescription } from './productResponseDtoDescription';
import type { ProductResponseDtoImageUrl } from './productResponseDtoImageUrl';
import type { ProductResponseDtoNotes } from './productResponseDtoNotes';
import type { ProductResponseDtoTotalStock } from './productResponseDtoTotalStock';
import type { ProductResponseDtoUnit } from './productResponseDtoUnit';
import type { VariantResponseDto } from './variantResponseDto';

export interface ProductResponseDto {
  id: string;
  sku: string;
  name: string;
  /** @nullable */
  description: ProductResponseDtoDescription;
  costPrice: number;
  sellingPrice: number;
  /** @nullable */
  unit: ProductResponseDtoUnit;
  trackInventory: boolean;
  isActive: boolean;
  hasVariants: boolean;
  /** @nullable */
  imageUrl: ProductResponseDtoImageUrl;
  /** @nullable */
  barcode: ProductResponseDtoBarcode;
  /** @nullable */
  notes: ProductResponseDtoNotes;
  category: CategoryResponseDto | null;
  brand: BrandResponseDto | null;
  manufacturer: ManufacturerResponseDto | null;
  /** Inventory info (null when hasVariants is true) */
  inventory: ProductInventoryDto | null;
  /**
   * Variants array (only present when hasVariants is true)
   * @nullable
   */
  variants: VariantResponseDto[] | null;
  /**
   * Total stock across all variants (only present when hasVariants is true)
   * @nullable
   */
  totalStock: ProductResponseDtoTotalStock;
  createdAt: string;
  updatedAt: string;
}